Started Writing Core Module

+++ Hydra/LibHydra.go
@@ -3,11 +3,14 @@ package Hydra
 import (
 	"Ariadne/CredManager"
 	"Ariadne/ElasticLog"
+	"context"
 	"sync"
 )
 
+const ModuleName = "HYDRA"
+
 type Module interface {
-	testCredential(string,string,string,*ElasticLog.Logger) bool
+	testCredential(string,string,string,*ElasticLog.Logger,context.Context) bool
 	preRunStart()
 	kill()
 	getModuleInfo()string
@@ -30,11 +33,13 @@ type LibHydra struct {
 	lock             sync.Mutex
 
 	Credentials chan CredManager.Cred
+	ctx 			 context.Context
+	cancelFunc 		 context.CancelFunc
 
 	module			 Module
 }
 
-func NewLibHydraModule(target string,credChannel chan CredManager.Cred,threads int,logger *ElasticLog.Logger,parentWaitGroup *sync.WaitGroup) *LibHydra {
+func NewLibHydraModule(target string,credChannel chan CredManager.Cred,threads int,logger *ElasticLog.Logger,parentWaitGroup *sync.WaitGroup,ctx context.Context) *LibHydra {
 	newLibHydraModule := &LibHydra{
 		target:          target,
 		Credentials:     credChannel,
@@ -42,6 +47,7 @@ func NewLibHydraModule(target string,credChannel chan CredManager.Cred,threads i
 		threads:         threads,
 		parentWaitGroup: parentWaitGroup,
 	}
+	newLibHydraModule.ctx, newLibHydraModule.cancelFunc = context.WithCancel(ctx)
 	return newLibHydraModule
 }
 
@@ -59,17 +65,21 @@ func (hydra *LibHydra) StartCracking(){
 }
 
 func (hydra *LibHydra) KillCrackingSession(){
+	if hydra.kill == true{
+		return
+	}
+	hydra.cancelFunc()
 	hydra.module.kill()
 	hydra.kill = true
 	close(hydra.Credentials) // to kill all threads
 }
 
 func (hydra *LibHydra) checkCredentials(credentials chan CredManager.Cred,group *sync.WaitGroup){
-
 	defer group.Done()
-	for credential := range credentials {
-		if !hydra.kill {
-			isValid := hydra.module.testCredential(hydra.target, credential.Username, credential.Password,hydra.logger)
+	select {
+	case credential := <- credentials:
+		if !hydra.kill{
+			isValid := hydra.module.testCredential(hydra.target, credential.Username, credential.Password,hydra.logger,hydra.ctx)
 			if !hydra.kill{
 				hydra.lock.Lock()
 				hydra.done += 1
@@ -77,17 +87,19 @@ func (hydra *LibHydra) checkCredentials(credentials chan CredManager.Cred,group 
 					hydra.KillCrackingSession() // when total is known - else (dynamic cases ex credtracker) channel needs to be closed
 				}
 				hydra.lock.Unlock()
-				hydra.logger.SendLog(ElasticLog.NewProgressLog(hydra.parentModuleName, hydra.target, hydra.done, hydra.Total))
+				hydra.logger.SendLog(ElasticLog.NewProgressLog(ModuleName, hydra.target,hydra.parentModuleName, hydra.done, hydra.Total))
 			}
 			if isValid && !hydra.kill{
 				hydra.foundCred = true
 				if hydra.findOneOnly {
 					hydra.KillCrackingSession()
 					hydra.kill = true // Update 2 - seems to work with different function to update the value in struct
-					hydra.logger.SendLog(ElasticLog.NewProgressLog(hydra.parentModuleName, hydra.target, hydra.Total, hydra.Total))
+					hydra.logger.SendLog(ElasticLog.NewProgressLog(ModuleName, hydra.target,hydra.parentModuleName, hydra.Total, hydra.Total))
 				}
 			}
 		}
+		case <- hydra.ctx.Done():
+			return
 	}
 
 }
